Add Spark batch job driver debug helper classes.
The class SparkBatchRemoteDebugJob and SparkBatchDebugSession
are introduced for getting necessary parameters from the
Spark cluster and creating SSH port forwarding tunnel.

The code is no referred now. And it will be used in Spark Job
IDE runner and executor.

Unit tests and integration tests with mocked http server
are included.

Signed-off-by: Zhang Wei <zhwe@microsoft.com>